==> scripts/3_mthds_6_glove_dict.R <==

# ---- 1. create functions ----

# function to create list of unique pos. & neg. words from list of seeds for
#   min. & max. pre-processed list of word_vector list
seeds2dicts <- function(word_vectors_ls, seeds_ls, max = 400) {

  words_df_ls <- lapply(word_vectors_ls, function(wv) {

    lapply(seeds_ls, function(seeds) {

      lapply(seeds, function(seed) {

        tryCatch(
          { # extract similar vectors (words) from seed vectors (words)
            word_vec <- wv[seed, , drop = FALSE]
            cos_sim  <- sim2(x = wv, y = word_vec, method = "cosine", norm = "l2")
            cos_sim_head <- head(sort(cos_sim[,1], decreasing = TRUE), max)

            word_sim_df <- data.frame(word = names(cos_sim_head),
                                      cos_sim = cos_sim_head,
                                      seed = seed,
                                      row.names = NULL)

            return(word_sim_df)
          },
          error = function(error_message) {
            message(seed)
            message(error_message)
            message()
            return(NULL)
          }
        )
      }) %>% bind_rows() # %>% na.omit() %>% distinct()
    })
  })

  dict_ls <- lapply(words_df_ls, function(words_df) {

    bind_rows(words_df, .id = "sentiment")
  })

  return(dict_ls)
}


# function to filter unique pos. or neg. words above given cosinus similarity in same amount
filter_words <- function(df, min_sim = .25) {

  # filter minimum of similarity
  filter(df, cos_sim > min_sim) %>%

    # keep unique words for each sentiment
    arrange(desc(cos_sim)) %>%
    distinct(sentiment, word, .keep_all = TRUE) %>%

    # keep words just in one sentiment
    group_by(word) %>%
    filter(n() == 1) %>%

    # keep just same amount of words for each sentiment
    group_by(sentiment) %>%
    mutate(max_size = max(row_number())) %>%
    ungroup() %>%
    mutate(max_size = min(max_size)) %>%
    group_by(sentiment) %>%
    filter(row_number() <= min(max_size))
}


# ---- 2. self-trained GloVe models with min & max pre-processing  ----
glove_ls <- readRDS("../data/glove_word_vectors.RDS")

load("../data/glove_seeds.Rdata")


# hand coded seeds
dict_hc_min <- seeds2dicts(list(onegram = glove_ls$onegram$min,
                                bigram  = glove_ls$bigram$min), seeds_hc_min)
dict_hc_max <- seeds2dicts(list(onegram = glove_ls$onegram$max,
                                bigram  = glove_ls$bigram$max),seeds_hc_max)

dict_hc <- list(min = lapply(dict_hc_min, filter_words),
                max = lapply(dict_hc_max, filter_words))


# rize and zorn seeds
dict_rz_min <- seeds2dicts(list(onegram = glove_ls$onegram$min,
                                bigram  = glove_ls$bigram$min), seeds_rz_min)
dict_rz_max <- seeds2dicts(list(onegram = glove_ls$onegram$max,
                                bigram  = glove_ls$bigram$max), seeds_rz_max)

dict_rz <- list(min = lapply(dict_rz_min, filter_words),
                max = lapply(dict_rz_max, filter_words))



# ---- 3. pre-trained german GloVe word vectors from deepset ----
wv_deepset <- vroom("C:/Users/munnes/Daten/frontiers_embeddings/deepset_vectors_glove.txt",
              delim = " ", col_names = FALSE)

row_names <- wv_deepset$X1
wv_deepset <- as.matrix(wv_deepset[, -1])
rownames(wv_deepset) = row_names

# 1309281 rows * 300 dimensions

# hand coded seeds
dict_glove_pre_hc <- seeds2dicts(list(wv_deepset), lapply(seeds_hc_min, tolower))

# rize and zorn seeds
dict_glove_pre_rz <- seeds2dicts(list(wv_deepset), lapply(seeds_rz_min, tolower))


dict_glove_pre <- lapply(c(0.3, 0.4, 0.5), function(similarity) {
  
  lapply(list(hc = dict_glove_pre_hc[[1]], 
              rz = dict_glove_pre_rz[[1]]), function(dict) {
                
     filter_words(dict, min_sim = similarity)
  })
})

names(dict_glove_pre) <- c("0.3", "0.4", "0.5")



# ---- 4. save all dictionarys ----
glove_dict_ls <- list(hc = dict_hc, rz = dict_rz, pre = dict_glove_pre)

saveRDS(glove_dict_ls, file = "../data/glove_dict_ls.RDS")
